Bootstrap: disable QFile::copy() due to lack of QTemporaryFile
Without QTemporaryFile, QFile::copy() creates a file with a fixed name
in the destination directory, meaning that it would step over another
thread or process trying to copy at the same time. That's dangerous for
buildsystem tools, given the prevalence of ninja -j / make -j
executions.

Nothing in the bootstrapped tools uses this function right now. This
commit will simply ensure nothing begins doing it in the future.

Change-Id: Icdcf51fa5d94618d622afffd94d1263cf9df181a
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>